requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
httpx
python-dotenv
google-genai
